pyproject.toml
[project]
name = "loom"
version = "0.1.0"
description = "Multi-agent swarm orchestrator for Claude Code and Gemini CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphiti-core>=0.28.2",
    "litellm>=1.83.2",
    "mcp>=1.27.0",
    "python-dotenv>=1.0.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-python>=0.25.0",
    "tree-sitter-typescript>=0.23.2",
    "tree-sitter-javascript>=0.23.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/loom"]

[project.optional-dependencies]
test = ["pytest>=8.0", "pytest-asyncio>=0.23"]
kan = ["torch>=2.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
markers = ["integration: requires Docker services (Neo4j, LiteLLM)"]
testpaths = ["tests"]
